run migrations before storing IDs in the cache
Previously the program tried to store some IDs provided via environment
variables in the database *before* running the database migrations.
While this was fine during development it prevents migrations from being
applied to a blank database, as storing the IDs errors out.

This commit switches the startup order to first run migrations and then
store IDs in the cache.