Fix really damn weird edge case
That might have triggered the endless "Ferris is typing" in #black-magic just now? IFFFF the stdout/stderr was somehow very long (~2000+ bytes) (it shouldn't be, with the input yandros gave it) AND the playground for some reason returns a bogus gist ID in its gist ID request which is also thousands of bytes long. If those conditions are met, reply_potentially_long_text calculates a negative number of leftover space which it tries to fit the response into. The negative number overflows into usize::MAX, so the bot counts down from usize::MAX downwards until it finds a value where text_body.is_char_boundary(index) is true. So essentially it would count down from usize::MAX to zero.

But I don't know if that was actually the cause there. It seems so unlikely that two impossible things would happen at the same time: that rust code giving a super long output when it shouldnt and doesnt when you retry it, AND the playground returning thousands of bytes when asking for the gist ID. Both those things not having happened before at all as far as I can tell.

Really weird

Well, even if that _wasn't_ the cause, _if_ those circumstances ever meet in the future, the bot will now not loop infinitely anymore. Hooray